# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "description": "Get all contacts",
                "operationId": "64796b9237ea222c5a3d303fbb744242",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contacts"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "description": "Store contact",
                "operationId": "23942d2fcfa7fbce9f161d2a0b5dffd0",
                "requestBody": {
                    "description": "Contact params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "description": "Show one contact",
                "operationId": "1f5a82f35f7c8dd39b6e9d6014c55b88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "description": "Update contact",
                "operationId": "b0864f1e1f7ef5a4ec304fa24cd3989c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Contact params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "description": "Delete contact",
                "operationId": "0f613c5c992734d816fe9e2cd5f05bf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/resource.json": {
            "get": {
                "summary": "/**",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contacts": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContactsResponse"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/PaginationResponse"
                    }
                },
                "type": "object"
            },
            "ContactsResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Contact"
                    }
                ]
            },
            "Contact": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "number"
                    },
                    "picture": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ContactResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Contact"
                    }
                ]
            },
            "CreateContactRequest": {
                "properties": {
                    "full_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "number"
                    },
                    "picture": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateContactRequest": {
                "properties": {
                    "full_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "number"
                    },
                    "picture": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationErrors": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "properties": {
                            "field": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "status_code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "status_code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginationResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Pagination"
                    }
                ]
            },
            "Pagination": {
                "properties": {
                    "total": {
                        "type": "integer"
                    },
                    "count": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "current_page": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}